feat(P09): TB-P09-T015 quantity invoice final gate

+++ src/backend/Host/Tooba.Host/Customer/CustomerPanelModels.cs
@@ -44,7 +44,7 @@ public sealed record CustomerOrderListItem(
     string Reference,
     DateTimeOffset SubmittedAt,
     int SellerCount,
-    decimal ItemCount,
+    int ItemCount,
     decimal PayableAmount,
     string Currency,
     string PaymentState,
